Move molgenis-data-index exceptions to their own package
