+++ molgenis-data-index/src/main/java/org/molgenis/data/index/exception/UnknownIndexException.java
@@ -1,4 +1,6 @@
-package org.molgenis.data.index;
+package org.molgenis.data.index.exception;
+
+import org.molgenis.data.index.exception.IndexException;
 
 import static java.util.Arrays.stream;
 import static java.util.stream.Collectors.joining;
